+++ internal/internal_event_handlers.go
@@ -153,6 +153,8 @@ type (
 		queryHandler    func(queryType string, queryArgs *commonpb.Payloads, header *commonpb.Header) (*commonpb.Payloads, error)
 		updateHandler   func(name string, id string, args *commonpb.Payloads, header *commonpb.Header, callbacks UpdateCallbacks)
 
+		cancellationReason string
+
 		logger                log.Logger
 		isReplay              bool // flag to indicate if workflow is in replay mode
 		enableLoggingInReplay bool // flag to indicate if workflow should enable logging in replay mode
@@ -414,15 +416,19 @@ func (wc *workflowEnvironmentImpl) Complete(result *commonpb.Payloads, err error
 	wc.completeHandler(result, err)
 }
 
-func (wc *workflowEnvironmentImpl) RequestCancelChildWorkflow(namespace string, workflowID string) {
-	// For cancellation of child workflow only, we do not use cancellation ID and run ID
-	wc.commandsHelper.requestCancelExternalWorkflowExecution(namespace, workflowID, "", "", true)
+func (wc *workflowEnvironmentImpl) GetCancellationReason() string {
+	return wc.cancellationReason
 }
 
-func (wc *workflowEnvironmentImpl) RequestCancelExternalWorkflow(namespace, workflowID, runID string, callback ResultHandler) {
+func (wc *workflowEnvironmentImpl) RequestCancelChildWorkflow(namespace, workflowID, reason string) {
+	// For cancellation of child workflow only, we do not use cancellation ID and run ID
+	wc.commandsHelper.requestCancelExternalWorkflowExecution(namespace, workflowID, "", "", reason, true)
+}
+
+func (wc *workflowEnvironmentImpl) RequestCancelExternalWorkflow(namespace, workflowID, runID, reason string, callback ResultHandler) {
 	// for cancellation of external workflow, we have to use cancellation ID and set isChildWorkflowOnly to false
 	cancellationID := wc.GenerateSequenceID()
-	command := wc.commandsHelper.requestCancelExternalWorkflowExecution(namespace, workflowID, runID, cancellationID, false)
+	command := wc.commandsHelper.requestCancelExternalWorkflowExecution(namespace, workflowID, runID, cancellationID, reason, false)
 	command.setData(&scheduledCancellation{callback: callback})
 }
 
@@ -1386,7 +1392,7 @@ func (weh *workflowExecutionEventHandlerImpl) ProcessEvent(
 		weh.commandsHelper.handleTimerCanceled(event.GetTimerCanceledEventAttributes().GetTimerId())
 
 	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CANCEL_REQUESTED:
-		weh.handleWorkflowExecutionCancelRequested()
+		weh.handleWorkflowExecutionCancelRequested(event)
 
 	case enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CANCELED:
 		// No Operation.
@@ -1689,7 +1695,8 @@ func (weh *workflowExecutionEventHandlerImpl) handleTimerFired(event *historypb.
 	timer.handle(nil, nil)
 }
 
-func (weh *workflowExecutionEventHandlerImpl) handleWorkflowExecutionCancelRequested() {
+func (weh *workflowExecutionEventHandlerImpl) handleWorkflowExecutionCancelRequested(event *historypb.HistoryEvent) {
+	weh.cancellationReason = event.GetWorkflowExecutionCancelRequestedEventAttributes().GetCause()
 	weh.cancelHandler()
 }